+++ docs/src/Preview.vue
@@ -48,7 +48,9 @@ const MAP = {
   toast: 'toast-201808191046.png',
   transition: 'transition-20180821.png',
   'tree-select': 'tree-select-201808092138.png',
-  checkbox: 'checkbox-20181110.jpeg'
+  checkbox: 'checkbox-20181110.jpeg',
+  rate: 'rate-20181120-1.png',
+  collapse: 'collapse-20181123.png'
 };
 
 export default {