+++ packages/search/index.ts
@@ -3,7 +3,7 @@ import { VantComponent } from '../common/component';
 VantComponent({
   field: true,
 
-  classes: ['cancel-class'],
+  classes: ['field-class', 'input-class', 'cancel-class'],
 
   props: {
     focus: Boolean,